+++ odra-macros/src/utils/expr.rs
@@ -119,15 +119,18 @@ pub fn serialized_length<T: ToTokens>(caller: &T) -> syn::Expr {
     parse_quote!(#caller.#fn_ident())
 }
 
+pub fn u8_serialized_len() -> syn::Expr {
+    parse_quote!(odra::casper_types::bytesrepr::U8_SERIALIZED_LENGTH)
+}
+
 pub fn failable_to_bytes<T: ToTokens>(caller: &T) -> syn::Expr {
     let fn_ident = super::ident::to_bytes();
     let ty = super::ty::to_bytes();
     parse_quote!(#ty::#fn_ident(&#caller)?)
 }
 
-pub fn to_bytes<T: ToTokens>(caller: &T) -> syn::Expr {
-    let fn_ident = super::ident::to_bytes();
-    parse_quote!(#caller.#fn_ident())
+pub fn serialize_enum() -> syn::Expr {
+    parse_quote!(Ok(odra::prelude::vec![self.clone() as u8]))
 }
 
 pub fn empty_vec() -> syn::Expr {